requirements.txt
pyserial
pyzmq==21.0.1